+++ src/commands/add.ts
