+++ src/commands/delete.ts
